TIP-458: Move YamlReader from BaseConnectorBundle to ConnectorComponent
