+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/readers.yml
@@ -5,8 +5,6 @@ parameters:
     pim_base_connector.reader.repository.group.class:         Pim\Bundle\BaseConnectorBundle\Reader\Repository\GroupReader
     pim_base_connector.reader.repository.variant_group.class: Pim\Bundle\BaseConnectorBundle\Reader\Repository\VariantGroupReader
 
-    pim_base_connector.reader.file.yaml.class:         Pim\Bundle\BaseConnectorBundle\Reader\File\YamlReader
-
     pim_base_connector.reader.cached.class:            Pim\Bundle\BaseConnectorBundle\Reader\CachedReader
 
 services:
@@ -52,26 +50,6 @@ services:
         class: '%pim_base_connector.reader.repository.variant_group.class%'
         arguments:
             - '@pim_catalog.repository.group'
-
-    pim_base_connector.reader.file.yaml:
-        class: '%pim_base_connector.reader.file.yaml.class%'
-
-    pim_base_connector.reader.file.yaml_multiple:
-        class: '%pim_base_connector.reader.file.yaml.class%'
-        arguments:
-            - true
-
-    pim_base_connector.reader.file.yaml_product:
-        class: '%pim_base_connector.reader.file.yaml.class%'
-        arguments:
-            - false
-            - false
-
-    pim_base_connector.reader.file.yaml_associations:
-        class: '%pim_base_connector.reader.file.yaml.class%'
-        arguments:
-            - false
-            - false
 
     # Deprecated: This service is deprecated and will be removed in 1.6
     pim_base_connector.reader.cached_association:
