+++ src/Pim/Bundle/InstallerBundle/Resources/config/fixture_loader.yml
@@ -16,7 +16,7 @@ services:
         class: '%pim_installer.fixture_loader.job_instances_builder.class%'
         arguments:
             - '@file_locator'
-            - '@pim_base_connector.reader.file.yaml'
+            - '@pim_connector.reader.file.yaml'
             - '@pim_base_connector.processor.job_instance'
             - [ '%pim_installer.fixture_loader.job_loader.config_file%' ]
